Finished Overhauling backend ExceptionHandler
Reduced boilerplate, simplified code, and introduced some new exception cases. Will add more like database exceptions, and general safety net exceptions.

+++ backend/src/main/java/com/ofs_160/webdev/Service/CustomerService.java
@@ -3,6 +3,7 @@ package com.ofs_160.webdev.Service;
 import com.ofs_160.webdev.ExceptionHandler.DuplicateCustomerException;
 import com.ofs_160.webdev.Model.Customer;
 import com.ofs_160.webdev.Repository.CustomerRepository;
+import com.ofs_160.webdev.ExceptionHandler.CustomerNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -23,11 +24,16 @@ public class CustomerService {
     }
 
     public Customer findCustomerById(int id) {
-        return customerRepository.findById(id).orElse(null);
+        return customerRepository.findById(id)
+            .orElseThrow(() -> new CustomerNotFoundException(Integer.toString(id)));
     }
 
     public void deleteCustomerById(int id) {
+        if (!customerRepository.existsById(id)) {
+            throw new CustomerNotFoundException(Integer.toString(id));
+        }
         customerRepository.deleteById(id);
+        
     }
 
     public void insertCustomer(Customer customer) {
@@ -35,6 +41,9 @@ public class CustomerService {
     }
 
     public void updateCustomer(Customer customer) {
+        if (!customerRepository.existsById(customer.getCustomer_id())) {
+            throw new CustomerNotFoundException(Integer.toString(customer.getCustomer_id()));
+        }
         customerRepository.save(customer);
     }
